Create Psalms Searcher 1
In part #1 I used the file bible-psalms.txt which contains all 2,461 verses of the Book of Psalms.

Each verse is on a separate line, with the reference first followed by the text of the particular verse.

For example, the first 3 lines of the file contain:

Psalms 1:1 Blessed is the man that walketh not in the counsel of the ungodly...
Psalms 1:2 But his delight is in the law of the LORD; and in his law doth he meditate...
Psalms 1:3 And he shall be like a tree planted by the rivers of water, that bringeth forth...


The Java program will ask the user for chapter and verse. For example:

What Book of Psalms chapter? 2
What Book of Psalms verse within the chapter? 6


The Java Program will print one of two outputs: the verse text without reference, or an appropriate message if the chapter and verse does not exist. 

For example (for inputs 2 & 6):

Yet have I set my king upon my holy hill of Zion.